completed moving dsu to kruskal
